decided on game, board, pieces & view as distinctions - now separate .dart files
