+++ docs/decorator/chess_game.dart
@@ -1,0 +1,121 @@
+import 'dart:html';
+import 'game.dart';
+import 'chess_pieces.dart';
+import 'chess_view.dart';
+import 'chess_board.dart';
+
+class ChessGame {
+  int turnCount = 0;
+  late GameBoard board = ChequeredBoard();
+  ChessPiece activePiece = EmptyPiece(0, 0);
+  late GameView view;
+  Element container;
+  bool rotated = false;
+
+  ChessGame(this.container) {
+    view = ChessBoardView(this, container);
+  }
+
+  void startGame() {
+    List<bool> initState = List.empty(growable: true);
+    initState.add(true);
+    initState.add(true);
+    initState.add(true);
+    initState.add(true);
+    initState.add(true);
+    initState.add(true);
+
+    setupPieces(initState);
+    refreshView();
+  }
+
+  String getTurnPlayer() {
+    return (turnCount % 2 == 0) ? "w" : "b";
+  }
+
+  void submitMove(int i, int j) {
+    if (validMove(activePiece, i, j)) {
+      movePiece(activePiece, i, j);
+      endTurn();
+      return;
+    }
+    clearMoveOptions();
+    activePiece = EmptyPiece(0, 0);
+
+    GamePiece piece = board.getPiece(i, j);
+
+    if (piece is ChessPiece) {
+      if (piece.colour == getTurnPlayer()) {
+        piece.move(board);
+        activePiece = piece;
+      }
+    }
+
+    refreshView();
+  }
+
+  void refreshView() {
+    view.displayBoard(board.getBoardState());
+    if (rotated) {
+      view.rotateBoard();
+    }
+  }
+
+  void clearMoveOptions() {
+    for (int i = 0; i < 8; i++) {
+      for (int j = 0; j < 8; j++) {
+        GamePiece piece = board.getPiece(i, j);
+        if (piece is ChessPiece) {
+          piece.threatened = false;
+        }
+      }
+    }
+  }
+
+  void endTurn() {
+    turnCount += 1;
+    activePiece = EmptyPiece(0, 0);
+    clearMoveOptions();
+    refreshView();
+  }
+
+  bool validMove(ChessPiece piece, int i, int j) {
+    GamePiece target = board.getPiece(i, j);
+    if (target is ChessPiece) {
+      return target.threatened;
+    }
+    return false;
+  }
+
+  void movePiece(ChessPiece piece, int i, int j) {
+    board.removePiece(piece.i, piece.j);
+    board.removePiece(i, j);
+    board.placePiece(piece, i, j);
+    piece.hasMoved = true;
+  }
+
+  void setupPieces(List<bool> state) {
+    board = ChequeredBoard();
+    if (state[0]) {
+      board = BoardWithPawns(board);
+    }
+    if (state[1]) {
+      board = BoardWithBishops(board);
+    }
+    if (state[2]) {
+      board = BoardWithKnights(board);
+    }
+    if (state[3]) {
+      board = BoardWithRooks(board);
+    }
+    if (state[4]) {
+      board = BoardWithQueens(board);
+    }
+    if (state[5]) {
+      board = BoardWithKings(board);
+    }
+
+    board.setupPieces();
+    refreshView();
+  }
+}
